Some other fixes

+++ src/controllers/movie-controller.js
@@ -2,6 +2,7 @@ import { Router } from "express";
 import movieService from "../services/movie-service.js";
 import castService from "../services/cast-service.js";
 import { isAuth } from "../middlewares/auth-middleware.js";
+import { getErrorMessage } from "../utils/errorUtils.js";
 
 const router = Router();
 
@@ -14,8 +15,14 @@ router.post('/create', isAuth, async (req, res) => {
     const movieData = req.body;
     const ownerId = req.user?._id;
 
-    // TODO: save movie data
-    await movieService.create(movieData, ownerId);
+    try {
+        await movieService.create(movieData, ownerId);
+    } catch (err) {
+        //Challenge: Show multi errors
+        const errorMessage = getErrorMessage(err);
+        
+        return res.render('movies/create', { error: errorMessage, movie: movieData });
+    }
 
     res.redirect('/');
 });
@@ -55,6 +62,14 @@ router.post('/:movieId/attach', isAuth, async (req, res) => {
 
 router.get('/:movieId/delete', isAuth, async (req, res) => {
     const movieId = req.params.movieId;
+
+    // Check if owner
+    const movie = await movieService.getOne(movieId);
+
+    if (movie.owner?.toString() !== req.user._id) {
+        res.setError('You cannot delete this movie!');
+        return res.redirect('/404')
+    }
 
     await movieService.remove(movieId);
